backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
openai
alembic
pytest
pytest-asyncio
httpx
python-docx
